Corrected package name of auth-service Refactored AuthService and AuthServiceImpl: - removed register method - added password coding when checking in login method
